server/requirements.txt
aiohttp==3.3.2
async-timeout==3.0.0
atomicwrites==1.1.5
attrs==18.1.0
blinker==1.4
Cerberus==1.2
chardet==3.0.4
click==6.7
coloredlogs==10.0
coverage==4.5.1
Flask==1.0.2
Flask-Cors==3.0.6
humanfriendly==4.16.1
idna==2.7
idna-ssl==1.1.0
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
more-itertools==4.2.0
multidict==4.3.1
pluggy==0.6.0
py==1.5.4
PyJWT==1.6.4
pymongo==3.7.1
pytest==3.6.3
raven==6.9.0
schematics==2.1.0
simplejson==3.16.0
six==1.11.0
waitress==1.1.0
Werkzeug==0.14.1
yarl==1.2.6